154fc67 Fix warnings showing on Windows dashboard
837f909 Another blight warning round
5270d60 Fix compilation issue
537a49e Fix a bunch of warning that was showing up on blight

+++ ParaViewCore/VTKExtensions/Default/vtkMaterialInterfaceIdList.cxx
@@ -56,7 +56,7 @@ void vtkMaterialInterfaceIdList::Initialize(vector<int> ids, bool preSorted)
 {
   // Prep.
   this->Clear();
-  int nLocalIds=ids.size();
+  size_t nLocalIds = ids.size();
   // Anything to do?
   if (nLocalIds<1)
   {
@@ -65,7 +65,7 @@ void vtkMaterialInterfaceIdList::Initialize(vector<int> ids, bool preSorted)
   // Make a copy of incoming, convert to list items to track
   // local id as items are sorted.
   this->IdList.resize(nLocalIds);
-  for (int localId=0; localId<nLocalIds; ++localId)
+  for (size_t localId=0; localId<nLocalIds; ++localId)
   {
     int globalId=ids[localId];
     this->IdList[localId].Initialize(localId,globalId);
@@ -86,7 +86,7 @@ int vtkMaterialInterfaceIdList::GetLocalId(int globalId)
          && this->IsInitialized);
 
   const int firstListItem=0;
-  const int lastListItem=this->IdList.size()-1;
+  const int lastListItem = static_cast<int>(this->IdList.size()) - 1;
   vtkMaterialInterfaceIdListItem itemToFind(globalId);
   return
     search(&this->IdList[0],firstListItem,lastListItem,itemToFind);
